requirements/requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
Django==4.2.6
ipython==8.16.1
psycopg2-binary==2.9.9
requests==2.31.0
Telethon==1.31.1
django-environ
gunicorn
django-celery-results
redis